requirements.txt
# Core dependencies
openai>=1.0.0
anthropic>=0.18.0
requests>=2.31.0
pydantic>=2.0.0
pyyaml>=6.0.0
python-dotenv>=1.0.0
click>=8.1.0
rich>=13.0.0

# Development dependencies
pytest>=7.4.0
black>=23.0.0
mypy>=1.5.0